Simplify components & type exports to just Controlled and Uncontrolled
