+++ source/Uncontrolled.tsx
@@ -1,17 +1,13 @@
 import React, { useState } from 'react'
-import Base, { BaseProps } from './Base'
+import Controlled, { ControlledProps } from './Controlled'
 
 // =============================================================================
 
 export type UncontrolledProps =
-  Omit<BaseProps, 'isZoomed' | 'onZoomChange'>
+  Omit<ControlledProps, 'isZoomed' | 'onZoomChange'>
 
-function Uncontrolled (props: UncontrolledProps) {
+export default function Uncontrolled (props: UncontrolledProps) {
   const [isZoomed, setIsZoomed] = useState(false)
 
-  return <Base {...props} isZoomed={isZoomed} onZoomChange={setIsZoomed} />
+  return <Controlled {...props} isZoomed={isZoomed} onZoomChange={setIsZoomed} />
 }
-
-export default typeof window !== 'undefined'
-  ? Uncontrolled
-  : (props: UncontrolledProps) => <>{props.children}</>
